Document compilation tracers in Code.Tracer
